Move encoding functions to encoding.c/h
