+++ include/tickle/hal.h
@@ -51,21 +51,6 @@ struct tt_Header;
 #include <tickle/hal_generic.h>
 #endif
 
-// Endian checking functions
-bool tt_is_native_endian(struct tt_Header* header);
-bool tt_is_reverse_endian(struct tt_Header* header);
-
-// Hash function
-uint32_t tt_hash_id(const char* type, const char* name);
-
-// Basic encoding/decoding utility functions
-void* tt_encode_buffer(void* buffer, uint32_t* tail, uint32_t len);
-void* tt_decode_buffer(void* buffer, uint32_t* head, uint32_t tail, uint32_t length);
-
-// String encoding/decoding functions
-bool tt_encode_string(void* buffer, uint32_t* tail, uint32_t buffer_size, const char* str);
-bool tt_decode_string(void* buffer, uint32_t* head, uint32_t tail, uint16_t* str_len, char** str);
-
 // Network functions
 int32_t tt_get_node_id();
 int32_t tt_bind(struct tt_Node* node);
